4 osa valmis, tehtävät 4.20-4.22 tehty

+++ osa4/blogilista/controllers/blog.js
@@ -4,13 +4,6 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 const { json } = require('express')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
 
 router.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -19,10 +12,9 @@ router.get('/', async (request, response) => {
 
 router.post('/', async (request, response) => {
     const body = request.body
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    if(!token || !decodedToken.id) {
+    if(!request.token || !decodedToken.id) {
         return response.status(401)-json({ error: 'token missing or invalid' })
     }
 
@@ -51,8 +43,16 @@ router.post('/', async (request, response) => {
 })
 
 router.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).end()
+    }
 })
 
 router.put('/:id', async (request, response) => {
